restrict methods and attributes
