+++ scheduler/scheduler-api/src/main/java/org/ow2/proactive/scheduler/common/job/factories/spi/model/ModelValidatorContext.java
@@ -28,23 +28,16 @@ package org.ow2.proactive.scheduler.common.job.factories.spi.model;
 import java.io.Serializable;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 import org.ow2.proactive.scheduler.common.job.JobVariable;
 import org.ow2.proactive.scheduler.common.job.TaskFlowJob;
+import org.ow2.proactive.scheduler.common.job.factories.spi.model.utils.RestrictedMethodResolver;
+import org.ow2.proactive.scheduler.common.job.factories.spi.model.utils.RestrictedPropertyAccessor;
+import org.ow2.proactive.scheduler.common.job.factories.spi.model.utils.RestrictedTypeLocator;
 import org.ow2.proactive.scheduler.common.task.Task;
 import org.ow2.proactive.scheduler.common.task.TaskVariable;
-import org.springframework.expression.EvaluationException;
-import org.springframework.expression.TypeLocator;
-import org.springframework.expression.spel.SpelEvaluationException;
-import org.springframework.expression.spel.SpelMessage;
 import org.springframework.expression.spel.support.StandardEvaluationContext;
-import org.springframework.util.ClassUtils;
-
-import com.google.common.collect.ImmutableSet;
 
 
 /**
@@ -66,6 +59,8 @@ public class ModelValidatorContext {
         spELVariables = new SpELVariables(variablesValues);
         spelContext = new StandardEvaluationContext(spELVariables);
         spelContext.setTypeLocator(new RestrictedTypeLocator());
+        spelContext.setMethodResolvers(Collections.singletonList(new RestrictedMethodResolver()));
+        spelContext.addPropertyAccessor(new RestrictedPropertyAccessor());
     }
 
     public ModelValidatorContext(Task task) {
@@ -105,116 +100,6 @@ public class ModelValidatorContext {
     public void updateTaskWithContext(Task task) {
         for (TaskVariable taskVariable : task.getVariables().values()) {
             taskVariable.setValue(spELVariables.getVariables().get(taskVariable.getName()).toString());
-        }
-    }
-
-    public static class RestrictedTypeLocator implements TypeLocator {
-
-        // A set of authorized types to prevent security breaches i.e executing maliciously code on the server
-        private final Set<String> authorizedTypes = ImmutableSet.of("String",
-                                                                    "java.lang.String",
-                                                                    "Integer",
-                                                                    "java.lang.Integer",
-                                                                    "Boolean",
-                                                                    "java.lang.Boolean",
-                                                                    "Double",
-                                                                    "java.lang.Double",
-                                                                    "Long",
-                                                                    "java.lang.Long",
-                                                                    "Float",
-                                                                    "java.lang.Float",
-                                                                    "Math",
-                                                                    "java.lang.Math",
-                                                                    "org.codehaus.jackson.map.ObjectMapper",
-                                                                    "ObjectMapper",
-                                                                    "javax.xml.parsers.DocumentBuilderFactory",
-                                                                    "DocumentBuilderFactory",
-                                                                    "java.io.StringReader",
-                                                                    "StringReader",
-                                                                    "org.xml.sax.InputSource",
-                                                                    "InputSource",
-                                                                    "org.json.simple.parser.JSONParser",
-                                                                    "JSONParser",
-                                                                    "Date",
-                                                                    "java.util.Date",
-                                                                    "ImmutableSet",
-                                                                    "com.google.common.collect.ImmutableSet",
-                                                                    "ImmutableMap",
-                                                                    "com.google.common.collect.ImmutableMap",
-                                                                    "ImmutableList",
-                                                                    "com.google.common.collect.ImmutableList");
-
-        private final ClassLoader classLoader;
-
-        private final List<String> knownPackagePrefixes = new LinkedList<String>();
-
-        public RestrictedTypeLocator() {
-            this(ClassUtils.getDefaultClassLoader());
-        }
-
-        public RestrictedTypeLocator(ClassLoader classLoader) {
-            this.classLoader = classLoader;
-            registerImport("java.lang");
-            registerImport("com.google.common.collect");
-            registerImport("org.codehaus.jackson.map");
-            registerImport("javax.xml.parsers");
-            registerImport("org.xml.sax.map");
-            registerImport("java.io");
-            registerImport("org.json.simple.parser");
-        }
-
-        /**
-         * Register a new import prefix that will be used when searching for unqualified types.
-         * Expected format is something like "java.lang".
-         * @param prefix the prefix to register
-         */
-        public void registerImport(String prefix) {
-            this.knownPackagePrefixes.add(prefix);
-        }
-
-        /**
-         * Remove that specified prefix from this locator's list of imports.
-         * @param prefix the prefix to remove
-         */
-        public void removeImport(String prefix) {
-            this.knownPackagePrefixes.remove(prefix);
-        }
-
-        /**
-         * Return a list of all the import prefixes registered with this StandardTypeLocator.
-         * @return a list of registered import prefixes
-         */
-        public List<String> getImportPrefixes() {
-            return Collections.unmodifiableList(this.knownPackagePrefixes);
-        }
-
-        /**
-         * Find a (possibly unqualified) type reference among a list of authorized types - first using the type name as-is,
-         * then trying any registered prefixes if the type name cannot be found.
-         * @param typeName the type to locate
-         * @return the class object for the type
-         * @throws EvaluationException if the type cannot be found
-         */
-        @Override
-        public Class<?> findType(String typeName) throws EvaluationException {
-            if (!authorizedTypes.contains(typeName)) {
-                throw new SpelEvaluationException(SpelMessage.TYPE_NOT_FOUND, typeName);
-            }
-            String nameToLookup = typeName;
-            try {
-                return ClassUtils.forName(nameToLookup, this.classLoader);
-            } catch (ClassNotFoundException ey) {
-                // try any registered prefixes before giving up
-            }
-            for (String prefix : this.knownPackagePrefixes) {
-                try {
-                    nameToLookup = prefix + "." + typeName;
-                    return ClassUtils.forName(nameToLookup, this.classLoader);
-                } catch (ClassNotFoundException ex) {
-                    // might be a different prefix
-                }
-            }
-            throw new SpelEvaluationException(SpelMessage.TYPE_NOT_FOUND, typeName);
         }
     }
