+++ scheduler/scheduler-api/src/main/java/org/ow2/proactive/scheduler/common/job/factories/spi/model/utils/RestrictedMethodResolver.java
@@ -1,0 +1,63 @@
+/*
+ * ProActive Parallel Suite(TM):
+ * The Open Source library for parallel and distributed
+ * Workflows & Scheduling, Orchestration, Cloud Automation
+ * and Big Data Analysis on Enterprise Grids & Clouds.
+ *
+ * Copyright (c) 2007 - 2017 ActiveEon
+ * Contact: contact@activeeon.com
+ *
+ * This library is free software: you can redistribute it and/or
+ * modify it under the terms of the GNU Affero General Public License
+ * as published by the Free Software Foundation: version 3 of
+ * the License.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ * If needed, contact us to obtain a release under GPL Version 2 or 3
+ * or a different license than the AGPL.
+ */
+package org.ow2.proactive.scheduler.common.job.factories.spi.model.utils;
+
+import java.util.List;
+import java.util.Set;
+
+import org.springframework.core.convert.TypeDescriptor;
+import org.springframework.expression.AccessException;
+import org.springframework.expression.EvaluationContext;
+import org.springframework.expression.MethodExecutor;
+import org.springframework.expression.spel.support.ReflectiveMethodResolver;
+
+import com.google.common.collect.ImmutableSet;
+
+
+/**
+ * @author ActiveEon Team
+ * @since 2019-06-07
+ */
+public class RestrictedMethodResolver extends ReflectiveMethodResolver {
+
+    private final Set<String> unAuthorizedMethods = ImmutableSet.of("getClass",
+                                                                    "notify",
+                                                                    "wait",
+                                                                    "notifyAll",
+                                                                    "finalize",
+                                                                    "clone");
+
+    @Override
+    public MethodExecutor resolve(EvaluationContext context, Object targetObject, String name,
+            List<TypeDescriptor> argumentTypes) throws AccessException {
+        if (unAuthorizedMethods.contains(name)) {
+            throw new AccessException("Unauthorized method: " + name);
+        }
+        return super.resolve(context, targetObject, name, argumentTypes);
+
+    }
+
+}
